refactor: reorganize project structure and rename files for clarity
